07_example_projects/03_alert_system/requirements.txt
paho-mqtt==1.6.1
pymongo==4.6.0
